Title: 최솟값 3개, Time: 173ms, Memory: 1MB, Status: Passed - CodeTree

+++ 240510/최솟값 3개/top-3-smallest-number.cpp
@@ -23,7 +23,7 @@ int main() {
             pq.pop();
             b = pq.top();
             pq.pop();
-            cout << a * b * pq.top() << '\n';
+            cout << (unsigned long long int)a * b * pq.top() << '\n';
             pq.push(b);
             pq.push(a);
         }
